Created tests for step1fun
I created tests folder and tests for step1fun in tests/testthat/test-step1fun. These tests check that the input items are the correct type and that their dimensions match. They also test that the output is the matrix and of the correct dimensions.
I also moved the README and LICENSE file to be inside of the gibbs.pack.
Finally, the tests folder for some reason went to 2022Spring-Gibbs-Sampler; I moved it to be inside of gibbs.pack.
Please let me know if anything I did here was incorrect so that I can fix it. (I'm not sure why the tests folder and the LICENSE file went to the outermost folder when my working directory was set to gibbs.pack)